fix(carta): int32 overflow in polygon clipping + increase clip buffer
Two interrelated bugs caused colored triangle artifacts at z16+:

1. Sutherland-Hodgman polygon clipper used int32 cross products for
   inside/outside tests. At z18, multipolygon coordinates can be 1.5M+
   extent units from tile origin; cross product 4128 * 1.5M = 6.2B
   overflows INT32_MAX. Fixed with int64_t casts in ct_clip_polygon,
   clip_ring_to_buffer, and ct_clip_linestring.

2. clip_buffer was 16 (reduced from 64 due to "precision errors" that
   were actually the int32 overflow). With only 16 units of overshoot,
   large polygon edges were clipped near the visible tile boundary,
   creating artificial straight edges visible as colored triangles.
   Increased to 1024 (~128px overshoot). Scanline rasterizer already
   clamps to [0,width]x[0,height] so no extra pixel cost.

+++ carta/src/ct_tile.c
@@ -5,6 +5,7 @@
 #include "ct_tile.h"
 #include "shared.h"
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <math.h>
 #include <pthread.h>
@@ -426,25 +427,27 @@ void ct_clip_linestring(const CTTilePoint *points, int num_points,
                 int code_out = code0 ? code0 : code1;
                 int x, y;
 
+                /* Use int64_t for intersection to prevent overflow with
+                 * distant coordinates (e.g., multipolygon edges at z18+). */
                 if (code_out & OUTCODE_TOP) {
                     int dy = y1 - y0;
                     if (dy == 0) break;  /* Degenerate: horizontal line at edge */
-                    x = x0 + (x1 - x0) * (max - y0) / dy;
+                    x = x0 + (int)((int64_t)(x1 - x0) * (max - y0) / dy);
                     y = max;
                 } else if (code_out & OUTCODE_BOTTOM) {
                     int dy = y1 - y0;
                     if (dy == 0) break;  /* Degenerate: horizontal line at edge */
-                    x = x0 + (x1 - x0) * (min - y0) / dy;
+                    x = x0 + (int)((int64_t)(x1 - x0) * (min - y0) / dy);
                     y = min;
                 } else if (code_out & OUTCODE_RIGHT) {
                     int dx = x1 - x0;
                     if (dx == 0) break;  /* Degenerate: vertical line at edge */
-                    y = y0 + (y1 - y0) * (max - x0) / dx;
+                    y = y0 + (int)((int64_t)(y1 - y0) * (max - x0) / dx);
                     x = max;
                 } else {
                     int dx = x1 - x0;
                     if (dx == 0) break;  /* Degenerate: vertical line at edge */
-                    y = y0 + (y1 - y0) * (min - x0) / dx;
+                    y = y0 + (int)((int64_t)(y1 - y0) * (min - x0) / dx);
                     x = min;
                 }
 
@@ -527,9 +530,11 @@ void ct_clip_polygon(const CTTilePoint *points, int num_points,
 
             /* Check if points are inside edge.
              * Cross product is negative when point is to the right of edge direction.
-             * Since edges go CCW, interior of rectangle is on the right (cross <= 0). */
-            int prev_inside = (x2 - x1) * (prev.y - y1) - (y2 - y1) * (prev.x - x1) <= 0;
-            int curr_inside = (x2 - x1) * (curr.y - y1) - (y2 - y1) * (curr.x - x1) <= 0;
+             * Since edges go CCW, interior of rectangle is on the right (cross <= 0).
+             * Use int64_t to prevent overflow: at z18, multipolygon coordinates can be
+             * 1.5M+ extent units from tile origin, and 4128 * 1.5M > INT32_MAX. */
+            int prev_inside = (int64_t)(x2 - x1) * (prev.y - y1) - (int64_t)(y2 - y1) * (prev.x - x1) <= 0;
+            int curr_inside = (int64_t)(x2 - x1) * (curr.y - y1) - (int64_t)(y2 - y1) * (curr.x - x1) <= 0;
 
             if (curr_inside) {
                 if (!prev_inside) {
@@ -664,9 +669,10 @@ static int clip_ring_to_buffer(const CTTilePoint *points, int ring_start, int ri
         for (int i = 0; i < input_count; i++) {
             CTTilePoint curr = input[i];
 
-            /* Check if points are inside edge */
-            int prev_inside = (x2 - x1) * (prev.y - y1) - (y2 - y1) * (prev.x - x1) <= 0;
-            int curr_inside = (x2 - x1) * (curr.y - y1) - (y2 - y1) * (curr.x - x1) <= 0;
+            /* Check if points are inside edge.
+             * Use int64_t to prevent overflow with distant multipolygon coordinates. */
+            int prev_inside = (int64_t)(x2 - x1) * (prev.y - y1) - (int64_t)(y2 - y1) * (prev.x - x1) <= 0;
+            int curr_inside = (int64_t)(x2 - x1) * (curr.y - y1) - (int64_t)(y2 - y1) * (curr.x - x1) <= 0;
 
             if (curr_inside) {
                 if (!prev_inside) {
